refactors to account for changes to Annotations class, fixed some imports
