+++ medacy/pipelines/clinical_pipeline.py
@@ -1,11 +1,11 @@
 import sklearn_crfsuite
 import spacy
 
-from medacy.pipelines.base import BasePipeline
+from medacy.pipelines.base.base_pipeline import BasePipeline
 from medacy.pipeline_components.tokenizers.clinical_tokenizer import ClinicalTokenizer
 from medacy.pipeline_components.feature_overlayers.gold_annotator_component import GoldAnnotatorComponent
-from medacy.pipeline_components.feature_overlayers.metamap import MetaMapComponent
-from medacy.pipeline_components.feature_overlayers.metamap import MetaMap
+from medacy.pipeline_components.feature_overlayers.metamap.metamap_component import MetaMapComponent
+from medacy.pipeline_components.feature_overlayers.metamap.metamap import MetaMap
 from medacy.pipeline_components.feature_extracters.discrete_feature_extractor import FeatureExtractor
