+++ medacy/pipelines/lstm_systematic_review_pipeline.py
@@ -1,6 +1,6 @@
 import spacy
 
-from medacy.pipelines.base import BasePipeline
+from medacy.pipelines.base.base_pipeline import BasePipeline
 from medacy.pipeline_components.feature_overlayers.gold_annotator_component import GoldAnnotatorComponent
 from medacy.pipeline_components.feature_extracters.discrete_feature_extractor import FeatureExtractor
 from medacy.pipeline_components.learners.bilstm_crf_learner import BiLstmCrfLearner
