+++ medacy/tests/pipelines/test_clinical_pipeline.py
@@ -1,6 +1,7 @@
 from unittest import TestCase
-from medacy.pipelines import ClinicalPipeline
-from medacy.pipeline_components import GoldAnnotatorComponent, MetaMap
+from medacy.pipelines.clinical_pipeline import ClinicalPipeline
+from medacy.pipeline_components.feature_overlayers.gold_annotator_component import GoldAnnotatorComponent
+from medacy.pipeline_components.feature_overlayers.metamap.metamap import MetaMap
 
 
 class TestClinicalPipeline(TestCase):
@@ -9,9 +10,7 @@ class TestClinicalPipeline(TestCase):
     def setUpClass(cls):
         metamap = MetaMap(metamap_path="/home/share/programs/metamap/2016/public_mm/bin/metamap",
                           cache_output=False)
-        cls.pipeline = ClinicalPipeline(metamap) # Will fail as MetaMap isn't installed
-
-
+        cls.pipeline = ClinicalPipeline(metamap)  # Will fail as MetaMap isn't installed
 
     def test_init(self):
         pipeline = self.pipeline
