+++ medacy/tests/tools/test_scores_counts_to_csv.py
@@ -5,6 +5,9 @@ from medacy.tools.scores_counts_to_csv import scores_and_counts_to_csv
 expected = """
 """
 
+
 class TestScoresCounts(unittest.TestCase):
 
     def test_scores_and_counts_to_csv(self):
+        pass
+
